requirements.txt
# AI Hadith Authenticator - Complete Requirements
# Core Flask Framework
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1

# Database Support
pymongo==4.6.0

# AI/ML Libraries
torch==2.1.1+cpu
transformers==5.5.3
numpy==1.24.3
gradio-client==0.10.1

# Hugging Face Integration
huggingface-hub==0.20.1
datasets==2.18.0

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
Pillow==10.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
pdf2image==1.17.0

# Date/Time
python-dateutil==2.8.2

# Email Support
blinker==1.7.0

# Development Tools
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.8

# Performance Monitoring
psutil==5.9.6

# Logging
colorlog==6.8.0

# JSON Processing
ujson==5.8.0

# Text Processing
nltk==3.8.1
regex==2023.10.21

# File Handling
python-magic==0.4.27

# Web Server (for production)
gunicorn==21.2.0
gevent==23.9.1

# SSL Support
certifi==2023.7.22

# API Documentation
flask-restx==1.3.0

# Rate Limiting
flask-limiter==3.5.0

# Session Management
Flask-Session==0.5.0

# CORS Support
flask-cors==4.0.0

# Template Engine
Jinja2==3.1.2

# Form Validation
wtforms==3.1.0
flask-wtf==1.2.1

# File Uploads
flask-uploads==0.2.1

# Caching
Flask-Caching==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring
sentry-sdk[flask]==1.38.0

# Configuration
pyyaml==6.0.1

# Testing
coverage==7.3.2
pytest-cov==4.1.0

# Code Quality
pylint==3.0.3
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Build Tools
setuptools==69.0.3
wheel==0.42.0

# Version Control
gitpython==3.1.40

# Deployment
docker==6.1.3
kubernetes==28.1.0
